backend/requirements.txt
# AI 주식 차트 분석기 - OpenAI API 백엔드

# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI API (최신 버전)
openai>=1.3.0

# 이미지 처리
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# HTTP 클라이언트
requests==2.31.0
aiohttp==3.9.1

# 환경 변수
python-dotenv==1.0.0

# 데이터 처리
pandas==2.1.3
python-dateutil==2.8.2

# 유틸리티
tqdm==4.66.1
loguru==0.7.2

# 개발 도구
pytest==7.4.3
black==23.11.0 